Ugly workaround for automatic updating
Map now loads all data points stored in database (can be occasionally dumped into observations.js manually if needed). This isn't great in terms of time, but I'm trying to do this all for free at the moment, and it seems like observation counts are bound to surpass 10,000 in no time.